+++ tests/infra/node.py
@@ -207,14 +207,14 @@ class Node:
                             "LedgerChunkRead"
                         )
 
-            # SnapshotTrigger operator feature is only supported from 7.0.0-dev14 onwards
+            # SnapshotCreate operator feature is only supported from 7.0.0-dev14 onwards
             if self.version is not None and Version(
                 strip_version(self.version)
             ) <= Version("7.0.0-dev13"):
                 if rpc_interface.enabled_operator_features:
-                    if "SnapshotTrigger" in rpc_interface.enabled_operator_features:
+                    if "SnapshotCreate" in rpc_interface.enabled_operator_features:
                         rpc_interface.enabled_operator_features.remove(
-                            "SnapshotTrigger"
+                            "SnapshotCreate"
                         )
 
     def __hash__(self):